+++ src/test/java/nextstep/subway/acceptance/line/LineAcceptanceTest.java
@@ -1,8 +1,8 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.line;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeEach;
+import nextstep.subway.acceptance.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
@@ -11,13 +11,10 @@ import org.springframework.http.MediaType;
 import java.util.HashMap;
 import java.util.Map;
 
-import static nextstep.subway.acceptance.AdministratorInfo.ADMIN_EMAIL;
-import static nextstep.subway.acceptance.AdministratorInfo.ADMIN_PASSWORD;
-import static nextstep.subway.acceptance.LineSteps.지하철_노선_목록_조회_요청;
-import static nextstep.subway.acceptance.LineSteps.지하철_노선_생성_요청;
-import static nextstep.subway.acceptance.LineSteps.지하철_노선_조회_요청;
-import static nextstep.subway.acceptance.MemberSteps.로그인_되어_있음;
-import static nextstep.subway.acceptance.RestAssuredStep.given;
+import static nextstep.subway.acceptance.line.LineSteps.지하철_노선_목록_조회_요청;
+import static nextstep.subway.acceptance.line.LineSteps.지하철_노선_생성_요청;
+import static nextstep.subway.acceptance.line.LineSteps.지하철_노선_조회_요청;
+import static nextstep.subway.utils.RestAssuredStep.given;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 노선 관리 기능")
