+++ src/test/java/nextstep/subway/acceptance/line/LineSectionAcceptanceTest.java
@@ -1,7 +1,8 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.line;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.subway.acceptance.AcceptanceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,19 +11,17 @@ import org.springframework.http.HttpStatus;
 import java.util.HashMap;
 import java.util.Map;
 
-import static nextstep.subway.acceptance.AdministratorInfo.ADMIN_EMAIL;
-import static nextstep.subway.acceptance.AdministratorInfo.ADMIN_PASSWORD;
-import static nextstep.subway.acceptance.LineSteps.*;
-import static nextstep.subway.acceptance.MemberSteps.로그인_되어_있음;
-import static nextstep.subway.acceptance.StationSteps.지하철역_생성_요청;
+import static nextstep.subway.acceptance.line.LineSteps.*;
+import static nextstep.subway.acceptance.station.StationSteps.지하철역_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 구간 관리 기능")
-class SectionAcceptanceTest extends AcceptanceTest {
+class LineSectionAcceptanceTest extends AcceptanceTest {
     private Long 신분당선;
 
     private Long 강남역;
     private Long 양재역;
+
     /**
      * Given 지하철역과 노선 생성을 요청 하고
      */
@@ -45,7 +44,7 @@ class SectionAcceptanceTest extends AcceptanceTest {
     void addLineSection() {
         // when
         Long 정자역 = 지하철역_생성_요청(관리자토큰,"정자역").jsonPath().getLong("id");
-        지하철_노선에_지하철_구간_생성_요청(관리자토큰, 신분당선, createSectionCreateParams(양재역, 정자역));
+        지하철_노선에_지하철_구간_생성_요청(관리자토큰,신분당선, createSectionCreateParams(양재역, 정자역));
 
         // then
         ExtractableResponse<Response> response = 지하철_노선_조회_요청(신분당선);
@@ -54,11 +53,42 @@ class SectionAcceptanceTest extends AcceptanceTest {
     }
 
     /**
+     * When 지하철 노선 가운데에 새로운 구간 추가를 요청 하면
+     * Then 노선에 새로운 구간이 추가된다
+     */
+    @DisplayName("지하철 노선 가운데에 구간을 추가")
+    @Test
+    void addLineSectionMiddle() {
+        // when
+        Long 정자역 = 지하철역_생성_요청(관리자토큰,"정자역").jsonPath().getLong("id");
+        지하철_노선에_지하철_구간_생성_요청(관리자토큰, 신분당선, createSectionCreateParams(강남역, 정자역));
+
+        // then
+        ExtractableResponse<Response> response = 지하철_노선_조회_요청(신분당선);
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+        assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(강남역, 정자역, 양재역);
+    }
+
+    /**
+     * When 지하철 노선에 이미 존재하는 구간 추가를 요청 하면
+     * Then 노선에 새로운 구간추가를 실패한다
+     */
+    @DisplayName("이미 존재하는 구간을 추가")
+    @Test
+    void addSectionAlreadyIncluded() {
+        // when
+        ExtractableResponse<Response> response = 지하철_노선에_지하철_구간_생성_요청(관리자토큰, 신분당선, createSectionCreateParams(강남역, 양재역));
+
+        // then
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+    }
+
+    /**
      * Given 지하철 노선에 새로운 구간 추가를 요청 하고
      * When 지하철 노선의 마지막 구간 제거를 요청 하면
      * Then 노선에 구간이 제거된다
      */
-    @DisplayName("지하철 노선에 구간을 제거")
+    @DisplayName("지하철 노선의 마지막 구간을 제거")
     @Test
     void removeLineSection() {
         // given
@@ -72,6 +102,27 @@ class SectionAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> response = 지하철_노선_조회_요청(신분당선);
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
         assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(강남역, 양재역);
+    }
+
+    /**
+     * Given 지하철 노선에 새로운 구간 추가를 요청 하고
+     * When 지하철 노선의 가운데 구간 제거를 요청 하면
+     * Then 노선에 구간이 제거된다
+     */
+    @DisplayName("지하철 노선의 가운데 구간을 제거")
+    @Test
+    void removeLineSectionInMiddle() {
+        // given
+        Long 정자역 = 지하철역_생성_요청(관리자토큰,"정자역").jsonPath().getLong("id");
+        지하철_노선에_지하철_구간_생성_요청(관리자토큰, 신분당선, createSectionCreateParams(양재역, 정자역));
+
+        // when
+        지하철_노선에_지하철_구간_제거_요청(관리자토큰, 신분당선, 양재역);
+
+        // then
+        ExtractableResponse<Response> response = 지하철_노선_조회_요청(신분당선);
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+        assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(강남역, 정자역);
     }
 
     private Map<String, String> createLineCreateParams(Long upStationId, Long downStationId) {
